LayoutTests/platform/chromium/fast/url/anchor-expected.txt

Test URLs that have an anchor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canonicalize('http://www.example.com/#hello, world') is 'http://www.example.com/#hello, world'
PASS canonicalize('http://www.example.com/#Â©') is 'http://www.example.com/#Â©'
PASS canonicalize('http://www.example.com/#𐌀ss') is 'http://www.example.com/#𐌀ss'
PASS canonicalize('http://www.example.com/#%41%a') is 'http://www.example.com/#%41%a'
PASS canonicalize('http://www.example.com/#\ud800\u597d') is 'http://www.example.com/#\uFFFD\u597D'
FAIL canonicalize('http://www.example.com/#a\uFDD0') should be http://www.example.com/#a﷐. Was http://www.example.com/#a�.
PASS canonicalize('http://www.example.com/#asdf#qwer') is 'http://www.example.com/#asdf#qwer'
PASS canonicalize('http://www.example.com/##asdf') is 'http://www.example.com/##asdf'
PASS canonicalize('http://www.example.com/#a\nb\rc\td') is 'http://www.example.com/#abcd'
PASS successfullyParsed is true

TEST COMPLETE
